requirements.txt
Flask==2.0.2
gunicorn==20.1.0
pymongo==3.12.1
python-dotenv==0.19.2
